+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@the-devoyage/graphql-accounts",
-  "version": "0.1.22",
+  "version": "0.2.0",
   "description": "",
   "main": "server.js",
   "scripts": {
@@ -16,28 +16,24 @@
     "@graphql-codegen/typescript": "^2.4.1",
     "@graphql-codegen/typescript-resolvers": "^2.4.2",
     "@types/bcryptjs": "^2.4.2",
-    "@types/is-empty": "^1.2.0",
-    "@types/validator": "^13.1.3",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
     "@typescript-eslint/parser": "^5.14.0",
     "eslint": "^8.10.0",
     "typescript": "^4.2.3"
   },
   "dependencies": {
-    "@apollo/federation": "^0.29.1",
+    "@apollo/subgraph": "^2.0.2",
     "@the-devoyage/mailer-connect": "^0.1.1",
     "@the-devoyage/micro-auth-helpers": "^0.1.20",
     "@the-devoyage/mongo-filter-generator": "^0.3.3",
-    "apollo-server": "^3.5.0",
+    "apollo-server": "^3.7.0",
     "bcryptjs": "^2.4.3",
     "dotenv": "^8.2.0",
-    "graphql": "^15.7.2",
-    "is-empty": "^1.2.0",
+    "graphql": "^16.5.0",
     "module-alias": "^2.2.2",
     "mongoose": "6.2.4",
     "nodemon": "^2.0.7",
-    "source-map-support": "^0.5.21",
-    "validator": "^13.5.2"
+    "source-map-support": "^0.5.21"
   },
   "_moduleAliases": {
     "@src": "dist/"
