+++ src/graphql/resolvers/resolvers.ts
@@ -4,5 +4,7 @@ import { Query } from "./query";
 import { Mutation } from "./mutation";
 
 export const resolvers: Resolvers = {
-  Account: { Account, Query, Mutation },
+  Account: { Account },
+  Query: { Query },
+  Mutation: { Mutation },
 };
