--- src/graphql/typeDefs/accounts/index.ts
@@ -1,1 +1,0 @@
-export * from "./accounts";
