+++ src/graphql/typeDefs/mutation/index.ts
@@ -1,0 +1,1 @@
+export * from "./mutation";
