+++ src/graphql/typeDefs/typeDefs.ts
@@ -1,3 +1,5 @@
-import { Account } from "./accounts";
+import { Account } from "./account";
+import { Query } from "./query";
+import { Mutation } from "./mutation";
 
-export const typeDefs = { Account };
+export const typeDefs = { Account, Query, Mutation };
