--- src/validate/accounts/index.ts
@@ -1,6 +1,0 @@
-export * from "./login";
-export * from "./register";
-export * from "./verify-email";
-export * from "./reset-password";
-export * from "./reset-activation-code";
-export * from "./update-email";
