--- src/validate/accounts/register/index.ts
@@ -1,1 +1,0 @@
-export * from "./register";
