--- src/validate/accounts/reset-activation-code/index.ts
@@ -1,1 +1,0 @@
-export * from "./reset-activation-code";
