--- src/validate/accounts/reset-password/index.ts
@@ -1,1 +1,0 @@
-export * from "./reset-password";
