--- src/validate/accounts/update-email/index.ts
@@ -1,1 +1,0 @@
-export * from "./update-email";
