--- src/validate/accounts/verify-email/index.ts
@@ -1,1 +1,0 @@
-export * from "./verify-email";
