--- src/validate/index.ts
@@ -1,1 +1,0 @@
-export * as Validate from "./validate";
