--- src/validate/validate.ts
@@ -1,1 +1,0 @@
-export * as Accounts from "./accounts";
